PostgresNode_Base is added
This base class requires:
 - explicit os_ops object

This base class declares the following abstract properties:
 - logs_dir
 - data_dir
 - utils_log_file
 - pg_log_file
